requirements.txt
asttokens==2.2.1
backcall==0.2.0
bidict==0.22.1
biosppy==1.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
Flask==2.2.3
fonttools==4.38.0
future==0.18.3
h5py==3.8.0
idna==3.4
ipykernel==6.21.2
ipython==8.11.0
iso8601==1.1.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kafka-python==2.0.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.0
matplotlib-inline==0.1.6
mitbih-processor==1.0.4
nest-asyncio==1.5.6
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pymongo==4.3.3
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
sendgrid==6.9.7
shortuuid==1.0.11
six==1.16.0
soundfile==0.11.0
stack-data==0.6.2
starkbank-ecdsa==2.2.0
threadpoolctl==3.1.0
tornado==6.2
traitlets==5.9.0
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.3
wfdb==4.1.0
